requirements.txt
PyYAML
pydantic
Jinja2
WeasyPrint
xhtml2pdf
Flask
click
